+++ src/M31.FluentApi.Generator/SourceAnalyzers/FluentApiDiagnostics.cs
@@ -32,6 +32,7 @@ internal static class FluentApiDiagnostics
         ConflictingControlAttributes.Descriptor,
         MissingBuilderStep.Descriptor,
         ReservedMethodName.Descriptor,
+        FluentLambdaMemberWithoutFluentApi.Descriptor,
     };
 
     internal static class MissingSetAccessor
@@ -129,7 +130,8 @@ internal static class FluentApiDiagnostics
         internal static readonly DiagnosticDescriptor Descriptor = new DiagnosticDescriptor(
             id: "M31FA006",
             title: "Several main attributes",
-            messageFormat: "Use only one of FluentMember, FluentPredicate, FluentCollection and FluentMethod",
+            messageFormat: "Use only one of FluentMember, FluentPredicate, FluentCollection, FluentLambda and " +
+                           "FluentMethod",
             category: "M31.Usage",
             defaultSeverity: DiagnosticSeverity.Error,
             isEnabledByDefault: true);
@@ -384,6 +386,25 @@ internal static class FluentApiDiagnostics
             Location location = attributeData.AttributeData.ApplicationSyntaxReference?
                 .GetSyntax().GetLocation() ?? Location.None;
             return Diagnostic.Create(Descriptor, location, methodName);
+        }
+    }
+
+    internal static class FluentLambdaMemberWithoutFluentApi
+    {
+        internal static readonly DiagnosticDescriptor Descriptor = new DiagnosticDescriptor(
+            id: "M31FA022",
+            title: "Fluent lambda member without Fluent API",
+            messageFormat: "FluentLambda can not be used on a member of type '{0}'. " +
+                           "Type '{0}' does not have a Fluent API.",
+            category: "M31.Usage",
+            defaultSeverity: DiagnosticSeverity.Error,
+            isEnabledByDefault: true);
+
+        internal static Diagnostic CreateDiagnostic(AttributeDataExtended attributeData, string type)
+        {
+            Location location = attributeData.AttributeData.ApplicationSyntaxReference?
+                .GetSyntax().GetLocation() ?? Location.None;
+            return Diagnostic.Create(Descriptor, location, type);
         }
     }
 }
