+++ src/M31.FluentApi.Tests/CodeGeneration/TestClasses/Abstract/PrivateFieldClass/Student.cs
@@ -1,5 +1,5 @@
-// Field is never used
-#pragma warning disable CS0169
+// Field is never assigned to, and will always have its default value
+#pragma warning disable CS0649
 
 using M31.FluentApi.Attributes;
